HTTP response parser handles reason line
