+++ libraries/DLHTTP/Test/DLHTTP.Test.mk
@@ -1,3 +1,3 @@
-SRC_FILES += DLUtility/DLUtility.Strings.cpp
+SRC_FILES += DLUtility/DLUtility.Strings.cpp DLHTTP/DLHTTP.Header.cpp DLHTTP/DLHTTP.RequestBuilder.cpp DLHTTP/DLHTTP.ResponseParser.cpp
 
 INC_DIRS += -IDLUtility/
